my-jetpack: refactoring -> create REST_Plans class
